refactor(purchase-order): 발주 store BE↔FE 매핑 헬퍼 분리
stores/purchaseOrder.js 가 506줄로 가이드라인을 초과했고, 매핑 헬퍼
(toFeOrder/toFeItem/toFeHistory + 카탈로그 매퍼 2개 + toBeCreateReq/
toBeUpdateReq) 가 store 본체와 같은 파일에 있어 state/getter/action
흐름을 가리고 있었다. 매핑 규칙은 BE 컨트랙트에 종속된 정적 변환이라
store 내부 reactive state 와 결이 다른 단위.

- stores/purchaseOrder/mappers.js 신설 — BE→FE 5개 + FE→BE 2개 +
  VENDOR_PROCESSING_STATUSES 상수 모두 이동
- 메인 store 는 import 후 state/getter/action 만 남겨 506→381줄로 축소
- 호출처(views/components 3곳) import 경로 변경 없음
- build 통과 확인
